index type conversion issues with BlasMatrix constructors

+++ trunk/linbox/tests/test-matrix-domain.C
@@ -1454,9 +1454,9 @@ int main (int argc, char **argv)
 {
 	bool pass = true;
 
-	static long n = 50;
-	static long m = 50;
-	static long k = 10;
+	static size_t n = 50;
+	static size_t m = 50;
+	static size_t k = 10;
 	static integer q = 2147483647U;
 	static int iterations = 1;
